requirements.txt
accelerate==0.30.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.43.1
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
av==14.0.1
blessed==1.20.0
braceexpand==0.1.7
certifi==2022.12.7
cffi==1.17.0
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.1
debugpy==1.8.2
decorator==4.4.2
decord==0.6.0
deepspeed==0.14.2
diffusers==0.29.2
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
email_validator==2.1.1
et-xmlfile==1.1.0
eval_type_backport==0.2.0
evaluate==0.4.3
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.13.1
flash-attn==2.7.2.post1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.3.1
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
gradio==3.23.0
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.4
ImageHash==4.3.1
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
linkify-it-py==2.0.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
nvidia-ml-py==12.550.52
omegaconf==2.3.0
openai==1.30.4
opencv-python==4.10.0.84
openpyxl==3.1.4
orjson==3.10.3
packaging==24.0
pandas==2.2.2
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
portalocker==2.10.1
prettytable==3.10.2
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
qwen-vl-utils==0.0.11
referencing==0.35.1
regex==2024.5.15
reportlab==4.2.0
requests==2.28.1
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.2.1
setproctitle==1.3.3
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.5.10
timm==1.0.7
tokenizers==0.19.1
toolz==0.12.1
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.0
triton==2.1.0
trl==0.8.6
typer==0.12.3
typing_extensions==4.12.2
tyro==0.8.4
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
unfoldNd==0.2.1
urllib3==1.26.13
uvicorn==0.29.0
uvloop==0.19.0
wandb==0.17.0
watchfiles==0.21.0
wcwidth==0.2.13
webdataset==0.2.96
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.2
